feat: let filenames in errors be relative to the current dir if possible (#5642)
# Description

## Problem

Resolves #5641

## Summary

It seems `FileMap`'s `name` function was the one that needed that
change.

To avoid doing a syscall to get the current directory every time a name
is needed, `FileMap` gets it at the beginning of the program. I think
this is fine as the current directory can't change while the compilation
is happening.

## Additional Context

None.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.