fix: allow using Self for function calls (#5629)
# Description

## Problem

Resolves #1791

## Summary

When resolving a path (any path), we now check if the first segment is
`Self` (a similar check is done in another places of our codebase). If
so, resolve the next path segments relative to that type. Only struct
types are supported here... in theory a trait type can also show up, but
that case is handled as a special case in the compiler. I'll create a
follow-up PR removing that special case and seeing if it works.

## Additional Context

None.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.