chore: Validate outputs in kernel circuits (#7706)
Kernel circuits have a validation step that checks their output, which
is constructed in an unconstrained function earlier. This check should
not be needed when running in unconstrained mode, so they were skipped.

However, this means that errors in the unconstraiend construction of the
output could go unnoticed in tests that run without real proofs enabled,
since those use the `-simulated` versions of kernels which are
unconstrained. To catch these errors, we now force kernel circuits to
always validate this output.

---------

Co-authored-by: Leila Wang <leizciw@gmail.com>