feat: add `std::meta::type_of` and `impl Eq for Type` (#5669)
# Description

## Problem

Part of #5668

## Summary

I originally wanted to just add `type_of`, but in order to test it in
some way I thought of comparing two types and checking whether they are
equal or not... which required implementing `Eq` for `Type`, which I
guess is useful any way.

## Additional Context

Another way could have been to be able to turn a `Type` into a string
and check that, but there's no unbounded `String` type at comptime...
Another way would be to `println` the type and check the program's
output, but there's no quick way to do it right now... but I think
asserting on equality is fine too, especially because the implementation
of `type_of` is very simple.

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.