feat: let LSP work will with code generated by macros (#5665)
# Description

## Problem

Code generated by macros (aztec_macros) ends up overlapping real code.
The reason this happens is that, for example, this code is injected into
a program:

```rust
let serialized_note = ...
```

The program above ends up with spans that are relative to that small
snippet (for example `serialized_note` ends up in bytes 4..=19) and that
might overlap some code that happens to be in that position in the same
file.

## Summary

In this PR we clear (set to a default of `0..0`) any spans that happen
in generated code, if running in LSP mode. The result is that generated
code doesn't overlap anymore with real code, and references, inlay
hints, go to definition, etc., start working fine.

There's no need to stop showing inlay hints inside contracts, as there's
no generated code that interferes there.

## Additional Context

Maybe in the near future aztec macros will not be needed anymore.
Implementing emptying spans is straight-forward so until that happens I
think it might be worth having these LSP features work well when macros
exist.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.