feat: Derive Eq & Default (#5667)
# Description

## Problem\*

Resolves #4594 

First traits to derive in the stdlib!

## Summary\*

Lets us derive the `Eq` and `Default` traits for almost any type.

I anticipate there will be issues with deriving generic types. I'll fix
this in another PR.

## Additional Context

Since derive is in the stdlib and we can derive traits now, I'm
arbitrarily marking the epic for metaprogramming as completed. There are
still bugs to fix and many helper functions we can add but the bulk of
the work and infrastructure is all in.

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.