fix: Don't store spans of quoted code (#5666)
# Description

## Problem\*

## Summary\*

Previously quoted code was stored as a `Vec<SpannedToken>` essentially
since that is what the lexer expects. This leads to issues when the code
is later inserted into another file however since the spans don't also
track the original file. Moreover, issuing an error for code when each
token may have a span originating from a completely different place is
confusing and difficult.

To fix this, I've just taken the Spans off the tokens and each time we
lex we have the caller location and use that as the same span for every
token. This means any failure to parse macro error will be issued at the
macro callsite instead of within the macro itself. I think this is
desired behavior although since it is an error message change it is not
really testable via CI.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.