fix: Fix evaluation of globals from other crates (#5658)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

We stored `comptime_scopes` in the Elaborator, but elaborators are
created one per each crate. So any globals from other crates would be
invisible to each next crate, or their value would be reset. I've moved
`comptime_scopes` to the NodeInterner to fix this.

## Additional Context

With this, `derive` now works in the stdlib. I've decided to split
adding that into a separate PR though.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.