fix: Print `comptime::Value` better (#5655)
# Description

## Problem\*

## Summary\*

Several Value variants are just ids which we previously couldn't print
well. I've added a helper type to hold a NodeInterner so that we can
print the names of these values as well. This will now print, e.g. `Eq`
instead of `(trait definition)`

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.