fix: Elaborate struct & trait annotations in the correct module (#5643)
# Description

## Problem\*

## Summary\*

Elaborates trait & struct annotations in the correct module.
We also save & restore type variables before elaborating items from the
interpreter. Otherwise this infers with checks that trait impls do not
overlap.

## Additional Context

This is necessary for `derive` to work in the stdlib. This will be
tested against regressions in the future as part of the PR adding
`derive` in the stdlib.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.