fix: workaround from_slice with nested slices (#5648)
# Description

## Problem\*

Resolves #5633 

## Summary\*
This a workaround and it does not solve the 2 underlying problems:
1. the no_predicate defer the inlining, so the compiler misses some
constant values. We should handle no_predicates in a better way. I
removed the failing assert_constant to workaround this.
2. the from_slice somehow flatten the slice of curve points and loose
the composite type of a point (field, field and bool). I worked around
this by not using from_slice.

## Additional Context
I have activated no-predicate for Pedersen in Noir. If the protocol
circuits work well with this version, we will finally be able to remove
the Pedersen black box!


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Maxim Vezenov <mvezenov@gmail.com>