fix: Fix where clause issue in items generated from attributes (#5673)
# Description

## Problem\*

Resolves #5671

## Summary\*

The issue was that we were adding trait constraints to the methods
themselves in `dc_mod` but this code isn't run for items generated from
attributes which only go through the elaborator. I've just moved this
code to the elaborator instead.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.