feat: Add a limited form of arithmetic on generics (#5625)
# Description

## Problem\*

Resolves #4784

## Summary\*

Adds a limited form of arithmetic operators on numeric generic types in
Noir. This includes only 5 operations: +, -, *, /, and %.

Going forward I'm going to distinguish between two definitions for
clearer communication between devs & users:
- Numeric generics: generics that use numeric type variables like `fn
foo<let N: u32>()`
- Arithmetic generics: when arithmetic operators are used on numeric
generics in a type position: `fn foo<let N: u32>() -> [Field; N + 1]`

## Additional Context

It is explicitly a _non-goal_ to always correctly return when two
_arithmetic generics_ are equal. For example, in our current type
checking we do not check whether applying the distributive law would
make two types equal. So you'd get a type error for trying to use `2*(N
+ 1)` where `2*N + 2` was expected. Users should generally consider the
type equality here to be largely structural and try to match that as
best as possible. ~~The only check performed besides simplification of
constant terms is commutativity for `+` and `*`. Even then, the check is
only one level deep. So `N * M * 3` is not equal to `3 * N * M`.~~
Update: we canonicalize commutative operations now, so we sort this and
it now type checks.

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.
- I'd like to have this in for some time before we start advertising it.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Maxim Vezenov <mvezenov@gmail.com>