feat: Derive `Ord` and `Hash` in the stdlib; add `std::meta::make_impl` helper (#5683)
# Description

## Problem\*

## Summary\*

`Ord` and `Hash` are the last two traits in the stdlib that can be
derived - and now we can.
I've also added `std::meta::make_impl` so that there's not so much
repeated code for each of this function. This also makes writing these
derive functions somewhat easier for users.

## Additional Context

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Michael J Klein <michaeljklein@users.noreply.github.com>