feat: add some `Module` comptime functions (#5684)
# Description

## Problem

Part of #5668

## Summary

Adds these comptime functions:
- `Quoted::as_module`
- `Module::name`
- `Module::is_contract`
- `Module::functions`
- `FunctionDefinition::name`

## Additional Context

Some notes:
- I made `Quoted::as_module` return `Option<ModuleId>`, while other
`Quoted::as_...` methods currently return a non-Option instead.
Eventually it would be nice to unify these but doing that in this PR
would have made it much bigger and complex, mainly because those methods
rely on methods that are used in many places in the compiler.
- For now I made `Module::name` and `FunctionDefinition::name` return
`Quoted`. Once we have a `String` type we could change them to `String`,
and the logic from going to Quoted -> String is very simple.

I was wondering... would a `String` type be what a Slice is for Arrays?

One more thought: I wonder if instead of returning `Option<ModuleId>` it
wouldn't be better to return some kind of `Result`. That way the caller
would know what the error is, if it happens. But I think we don't have
`Result` yet in Noir, right?

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>