[chore] hackily apply no_predicates to call creation (#7804)
Ideally we'd do #7729, but as noted there that is currently not working.
The noir team is looking into improving how `no_predicates` work to
eventually solve this. As a temporary measure however, wrapping the
construction of the call in a function that has the attribute gets us
the expected gate count reduction.