feat: Hook up secondary calldata column in dsl (#7759)
Previously we could use a single calldata and return_data from noir with
support from the bberg backend. With
[this](https://github.com/noir-lang/noir/pull/5599/files) PR, noir has
an interface for multiple calldata entities. The backend has support for
two calldata columns (`calldata` and `secondary_calldata`). This work
hooks up a second calldata column in dsl.

The main limitation of this work is that there is no way to distinguish
between the two calldata columns in dsl. This is OK for the operations
within a single circuit because in that context there is no important
distinction between the two calldata columns (`calldata`,
`secondary_calldata`). It does cause a problem however in the mechanism
for linking two circuits via the databus. This is because we need to
know which calldata corresponds to app data and which corresponds to
previous kernel data in order to prove that the connection was made
faithfully. The ideal solution is probably to treat `secondary_calldata`
(possibly rename to `app_calldata`?) as a unique entity in noir (similar
to how `calldata` and `return_data` are treated as different entities),
rather than allowing arbitrarily many individual `calldata` entities. I
made an issue
[here](https://github.com/AztecProtocol/barretenberg/issues/1070).

---------

Co-authored-by: sirasistant <sirasistant@gmail.com>