feat: Add array_to_str_lossy (#5613)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

I noticed we had a way to convert from strings to byte arrays but no way
to do the reverse.

## Additional Context

Draft because I'm unsure how to implement this in acir-gen where we
can't know the actual value of the array.

## Documentation\*

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.