feat: split merge into recursive verification and proving (#7801)
ClientIvc/AztecIvc originally had a single method `accumulate(circuit)`
which did two things under the hood: 1) append recursive PG/merge
verifiers, and 2) perform the prover work for PG/merge. To facilitate
the new noir-bb interface that Cody and I have designed, we need to be
able to perform these two operations separately. This PR sets the stage
for that by re-expressing the `accumulate()` method of AztecIvc in terms
of two new methods: `complete_kernel_circuit_logic()` and
`execute_accumulation_prover()`.

The bulk of the work in doing this was similarly splitting up the method
`goblin.merge()` which performs both recursive verification for the
previous merge proof and merge proving for the circuit being
accumulated. We now have individual methods `goblin.prove_merge()` and
`goblin.verify_merge()` (which have been integrated into AztecIvc only).
Merge proofs are added to a `merge_verification_queue` as they are
created, which occurs on every accumulation round. When we get to a
kernel we recursively verify each proof then empty the queue. On the
first kernel there is a single merge proof to verify and in subsequent
kernels there are two.

closes https://github.com/AztecProtocol/barretenberg/issues/1063