feat: add mutating FunctionDefinition functions (#5685)
# Description

## Problem

Part of #5668

## Summary

## Additional Context

I think we are missing a `FunctionDefinition::body`, but that could come
later.

I had to get a `FuncMeta` many times in some functions because Rust
wouldn't let me get two mutable borrows of interner.

Generics handling is missing... but I think right now generics can't
appear in these functions, as attributes only run on top-level
functions, not on impl or trait impl functions, but let me know if this
is not the case.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.