fix(plonky2): fix generating Verifier.toml bug
Fixed the bug of mismatching data in the generated Verifier.toml file
when using `nargo prove`. This was caused by an incorrect matching of
the input data with the values in WitnessMap. More specifically the bug
was caused by calling the `decode` method on the object returned by the
function `public_abi` which filters needed parameters for the correct
decoding process.

The fix consists of the change of not calling the `public_abi` before
`decode` and passing a public_abi object to `write_inputs_to_file` which
only writes the parameters in the given abi object into Verifier.toml.
All those parameters are public.

Moved the bug related test into expected success folder.