chore!: Sequencer no longer proves (#7860)
Removes the code path where the sequencer would submit proofs for a
given block, so this reponsibility is now exclusive of the prover-node.
This allows us to remove the prover instance from the aztec node
altogether, and drop the `PROVER_DISABLED` and `SEQ_SKIP_SUBMIT_PROOFS`
env vars since are no longer needed.

Since the sequencer still needs to submit block roots as it is
publishing tx effects instead of tx objects, we keep using the
orchestrator as a means to assemble an L2Block using simulated circuits,
relying on a block simulator backed by the `TestCircuitProver`. We
should be able to remove this altogether once we migrate to tx objects.