fix: unexpose get note nonces on pxe (#7889)
This PR unexposes `getNoteNonces` on the PXE by making notes returned
from `getIncomingNotes` and `getOutgoingNotes` include a nonce.

This is done by extending `ExtendedNote` to create
`ExtendedNoteWithNonce`, a new type that adds a required property,
nonce, to the extended note.

I chose this approach instead of another e.g. adding an optional
property on `ExtendedNote` because these data structures are
fundamentally different—used in different places, for different reasons.
It didn't make sense to me to try and shoehorn both uses to one struct,
for the gain of slight de-duplication.

Because `ExtendedNote` is only used for notes added to or retrieved from
the PXE, there were not that many required changes.

Note: I'm definitely not a fan with the naming but cannot currently
think of anything better. Open to suggestions from the reviewer.

---------

Co-authored-by: Jan Beneš <janbenes1234@gmail.com>