feat: add `Type::get_trait_impl` (#5716)
# Description

## Problem

Part of #5668

## Summary


## Additional Context

I thought about adding some `TraitImpl` methods but given that this
introduces a new type and a new value I wanted to get feedback on that
approach first. There's also a TODO that needs to be addressed.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.