fix: Replace unused ArrayGet/Set with constrain if possibly out of bounds (#5691)
# Description

## Problem

For #5296

## Summary

In DIE, when we find an `ArrayGet/Set` that we could remove, if that
instruction might be an index out of bounds we still remove it but leave
a constrain check too.

## Additional Context

For now the code is a bit messy, I just wanted to get it working first.

Also, it's not working for slices yet because we can't insert out of
bounds checks if we don't know the slice length. For that we'll need to
add the checks like we do for `ArrayGet` for slice (this could be done
in a separate PR, though).

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.