fix: lay plumbing for having simulations throw an error if they cannot be added in a block (#7839)
This PR gives an option to make a simulation of a tx fail, if the tx
cannot be included in a block and added to the state.

To minimize impact on callsites / disruption, this update does not make
this validation necessary, and defaults to skipping tx validation,
exhibiting previous non-validated behavior. This will be changed to be
default behavior in the future.

e.g. If a simulation produces duplicate nullifiers, or nullifiers that
already exist in a state tree, the results of this simulation should not
be returned, but should warn users that the transaction simulated is
impossible to actually be added to a block due to being an invalid
transaction.

The method for achieving the above is that a new API on the node was
created, used for validating the correctness of the metadata and
side-effects produced by a transaction. A transaction is deemed valid if
and only if the transaction can be added to a block that can be used to
advance state.

Another note just for reference: there is weirdness in e2e_prover, that
fails the proof validation.

---------

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>