feat: add `unsafe` blocks for calling unconstrained code from constrained functions (#4429)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This is an experimental PR to play around with unsafe blocks as a method
to make it clearer when unconstrained values are entering the circuit to
avoid under-constrained errors.

Calls to unconstrained functions will now fail to compile unless inside
of an `unsafe` block. Note that any code inside of an `unsafe` block
still lays down constraints as usual so it's possible to fully constrain
the return values from any unconstrained function calls before returning
a safe value out from the block.


## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Ary Borenszweig <asterite@gmail.com>
Co-authored-by: jfecher <jake@aztecprotocol.com>