fix(frontend): Continue type check if we are missing an unsafe block (#5720)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

No issue as found quickly and want to push a patch ASAP

## Summary\*

Removed the `return Type::Error` when we run into a missing unsafe
block. Also switched the call to `lookup_function_from_expr` inside
`type_check_call` to use an if let as this was causing panics when
checking higher order functions that come from params.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.